fixing warning

+++ plugins/mtoa/utils/HashUtils.h
@@ -4,5 +4,5 @@
 
 inline int getHash(AtNode* node)
 {
-   return AiNodeGetStr(node, AtString("name")).hash();
+   return (int)AiNodeGetStr(node, AtString("name")).hash();
 }
